Move the remaining classes to their own modules.
